DBZ-8826 Refactor tests for temporal type string
